Branch: feature-Gtest, 1. make video-test alone, 2. do some small change of cmakeLists.txt
