+++ CMakeLists.txt
@@ -33,4 +33,3 @@ fetch_googletest(
 enable_testing()
 add_subdirectory(test)
 
-
