+++ README.md
@@ -4,7 +4,7 @@
 
 分支管理：
 
-**dev/feature** : 开发分支，最新的功能、优化等统一合入此分支(简称dev)，前身是test
+**dev** : 开发分支，最新的功能、优化等统一合入此分支(简称dev)，前身是test
 
 **alpha/v01** : 经集成测试后确认稳定的分支(可以展示的分支)，v01,v02代表版本号，每次迭代在dev分支测试后拉出一个新alpha分支，例如本周日完成一次迭代，三个人一起对dev分支确认测试通过后拉出一个新分支alpha/v02，并将这次迭代新增的功能、优化等添加到开发记录中。例如：
 
@@ -15,6 +15,14 @@ alpha/v01 2020/10/22 识别EAN13
 
 
 **master**: 经大量测试及代码规范检查之后确认可以合入opencv的分支，由稳定的alpha版本合入。
+
+
+
+规范commit信息格式：
+
+branch xxx
+
+commit 内容说明
