+++ test/main.cpp
@@ -1,4 +1,12 @@
-#include "gtest/gtest.h"
+/*
+ * @Author: your name
+ * @Date: 2020-11-04 20:55:54
+ * @LastEditTime: 2020-11-04 21:15:34
+ * @LastEditors: your name
+ * @Description: In User Settings Edit
+ * @FilePath: \Barcode\test\main.cpp
+ */
+#include "test_precomp.hpp"
 
 int main(int argc, char **argv) {
     std::cout << __DATE__ << " " << __TIME__ << std::endl;
